Add project memberships verbs
This is following the project membership api endpoint: https://www.pivotaltracker.com/help/api/rest/v5#Project_Memberships

My usecase was to retrieve the person who post the story as a Person object instead of only requester by id.

There is no endpoint to find a person on the api, project memberships was the closest endpoint to achieve this goal. It will give also all membership (With Personobject inside) who are in the project.